requirements.txt
asgiref==3.8.1
coverage==7.5.4
Django==5.0.6
django-guardian==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
npm==0.1.1
optional-django==0.1.0
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
